+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   Playground in Firefox private browsing mode, which does not support service
   workers (https://bugzilla.mozilla.org/show_bug.cgi?id=1320796).
 
+- Added getters for code-editor to access the cursor position, and the tokens under cursor
+
 ### Changed
 
 - Upgraded dependencies, including CodeMirror.